refactor(web): split Group 4 (external-API) routes out of create_app()
Extracts weather/astro/radar, music, gcal, email, voice/tts, vision, stt,
scan/receipt, and location/map routes from app.py's create_app() into
per-domain APIRouter modules under interfaces/web/routes/, following the
Phase 6b Group 1-3 pattern. No path/method/response changes; verified via
an openapi()["paths"] diff against origin/main (identical) and the full
test suite (186 passed).
